requirements.txt
openai
python-dotenv
streamlit
streamlit-chat